pyproject.toml
[tool.poetry]
name = "image-codec"
version = "0.1.0"
description = ""
authors = ["mcdnmd"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.0"
pillow = "^10.3.0"
pytest = "^8.2.0"
torchvision = "^0.18.0"
numpy = "^1.26.4"
pybind11 = "^2.12.0"
setuptools = "^69.5.1"
matplotlib = "^3.8.4"
tensorflow = "^2.16.1"
tqdm = "^4.66.4"
wandb = "^0.17.0"
pydantic = "^2.7.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
